Solve x86 compilation woes with size_t

+++ src/cascadia/TerminalApp/fzf/fzf.h
@@ -7,8 +7,8 @@ namespace fzf::matcher
 {
     struct TextRun
     {
-        int32_t Start;
-        int32_t End;
+        size_t Start;
+        size_t End;
     };
 
     struct MatchResult
